Made Application more convenient to use. Constructor assigns the current_application* to itself
Opted for function pointers for the bootstrap and tick methods so that run() only needs implemented once in the base class.

I suggest making static methods in your inherited type and just assigning the function pointers to the implementations in your constructor, then, when creating an instance, just call run(). Fire and forget!